+++ cejug-classifieds-server/src/net/java/dev/cejug/classifieds/server/ejb3/bean/interfaces/CRUDLocal.java
@@ -54,5 +54,5 @@ public interface CRUDLocal<T> {
 	/**
 	 * TODO: to comment.
 	 */
-	ServiceStatus delete(final int id);
+	ServiceStatus delete(final Long id);
 }
